requirements.txt
PyMySQL==1.0.2
PyYAML==6.0
redis==3.5.3

